Addition of ARM A64 vistior

+++ src/main/java/com/github/ahhoefel/lang/ast/visitor/FormatVisitor.java
@@ -108,7 +108,7 @@ public class FormatVisitor implements Visitor {
 
     @Override
     public void visit(IntegerLiteralExpression expr, Object... objs) {
-        out.add(Integer.toString(expr.getValue()));
+        out.add(Long.toString(expr.getValue()));
     }
 
     @Override
